requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
numpy==1.26.2
xlsxwriter==3.1.9
gunicorn==21.2.0
urllib3==2.1.0
